requirements.txt
ultralytics>=8.0.0
gradio>=4.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
PyYAML>=6.0
